// src/shared/api/generated/types/PublicSiteSettingsResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const publicSiteSettingsResponseDtoHomeSectionsOrderEnum = {
  "home.quick-access": "home.quick-access",
  "home.resources": "home.resources",
  "home.gia-reference": "home.gia-reference",
  "home.official-resources": "home.official-resources",
} as const;

export type PublicSiteSettingsResponseDtoHomeSectionsOrderEnumKey =
  (typeof publicSiteSettingsResponseDtoHomeSectionsOrderEnum)[keyof typeof publicSiteSettingsResponseDtoHomeSectionsOrderEnum];

export type PublicSiteSettingsResponseDto = {
  /**
   * @type string
   */
  giaHotlinePhone: string;
  /**
   * @type string
   */
  informationPhone: string;
  /**
   * @type string
   */
  egeTrustPhone: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type array
   */
  homeSectionsOrder: PublicSiteSettingsResponseDtoHomeSectionsOrderEnumKey[];
};